refactor(admin): extract shared CMS draft state into AdminCmsStateService
Move the self-contained CmsDraftManager engine and the shared CMS
draft/autosave registries (home sections, page builders, blog posts)
out of the 16k-line AdminComponent into a dedicated, component-provided
AdminCmsStateService.

- New content/admin-cms-state.service.ts holds cmsHomeDraft, cmsPageDrafts,
  cmsBlogDrafts and the ensurePageDraft/ensureBlogDraft factories;
  CmsDraftManager is moved verbatim and exported.
- AdminComponent injects the service (component-scoped provider to preserve
  the previous per-instance draft lifetime) and delegates via getters and
  methods, so all existing call sites and the (component as any) test access
  are unchanged. Behaviour-preserving: no logic changed.
- Domain types consumed by the service are exported from admin.component.ts;
  the service imports them type-only, so there is no runtime import cycle.
- Adds admin-cms-state.service.spec.ts (behavioural coverage of the
  registries plus the moved draft/autosave engine).

R5 drag-active and R6 CMS aria-announce state are untouched.
